Add functionality to ignore .env file from auto-generation with #noexample flag
This commit introduces a new feature that enables developers to opt out of the automated .env.example file generation for a specific .env file by adding a `#noexample` flag on the first line. This flag can be added in any casing and the system will ignore such .env files for auto-generation. Furthermore, if the flag is added after an .env.example file has been generated, the system will automatically delete the existing .env.example file and stop further updates until the flag is removed.