Added exclusion mechanism for .env files in .env.example generation
This commit introduces a new functionality that allows developers to exclude certain .env files from being processed by the EnvAutoExample extension. This is achieved by adding a #noexample flag to the first line of the .env file that should be ignored. This update provides users with greater control and flexibility when dealing with different .env files in their projects.